% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-facet-borders.R
\name{add_facet_borders}
\alias{add_facet_borders}
\title{Adds borders around panels and strips}
\usage{
add_facet_borders(color = "grey70", colour)
}
\arguments{
\item{color, colour}{The color of the border.}
}
\description{
This is a convenience function to add borders around panels and strips in
ggplot2. All it does it modify the panel.border and strip.background theme
elements to have the same specified color.
Credit to the \code{cowplot} package back for the idea. I only added the extra
strip.background border because I think it looks better.
}
\examples{
library(ggplot2)
library(dplyr)
library(palmerpenguins)

# Load the default font (Roboto Condensed)
extrafont::loadfonts(quiet = TRUE)

p <- penguins \%>\%
  filter(!is.na(bill_length_mm)) \%>\%
  ggplot(aes(x = bill_length_mm, y = bill_depth_mm, color = species)) +
  geom_point(aes(shape = species), size = 3, alpha = 0.7) +
  geom_smooth(method = "lm", formula = "y ~ x", se = FALSE) +
  scale_color_brewer(palette = "Set1") +
  labs(
    title = "Penguin bill dimensions",
    subtitle = "Bill length and depth for different penguin species",
    x = "Bill length (mm)", y = "Bill depth (mm)",
    color = "Penguin species", shape = "Penguin species",
    caption = "Data from the palmerpenguins package"
  ) +
  facet_wrap(~species, nrow = 1)

p + theme_td() + add_facet_borders()
}
