% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distill.R
\name{create_post_tdunn}
\alias{create_post_tdunn}
\title{Create a new post with the tdunn.ca template}
\usage{
create_post_tdunn(..., open = TRUE)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[distill:create_post]{distill::create_post}}
  \describe{
    \item{\code{title}}{Post title}
    \item{\code{collection}}{Collection to create the post within (defaults to "posts")}
    \item{\code{author}}{Post author. Automatically drawn from previous post if not provided.}
    \item{\code{slug}}{Post slug (directory name). Automatically computed from title if not
provided.}
    \item{\code{date}}{Post date (defaults to current date)}
    \item{\code{date_prefix}}{Date prefix for post slug (preserves chronological order for posts
within the filesystem). Pass \code{NULL} for no date prefix.}
    \item{\code{draft}}{Mark the post as a \code{draft} (don't include it in the article listing).}
    \item{\code{edit}}{Open the post in an editor after creating it.}
  }}

\item{open}{Logcal. If \code{TRUE} (default), opens the created file.}
}
\description{
This function wraps \code{\link[distill:create_post]{distill::create_post()}} to create a template post for my
website https://tdunn.ca.
}
\details{
Credit goes to Erik Ekholm:
EE (2021, April 5). Eric Ekholm: Personalizing the Distill Template.
Retrieved from
https://www.ericekholm.com/posts/2021-04-02-personalizing-the-distill-template/
}
\examples{
create_post_tdunn("Post title", date = Sys.Date(), date_prefix = TRUE,
                  draft = TRUE, open = FALSE)
}
