% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunn.test.R
\name{dunn.test}
\alias{dunn.test}
\alias{dunn.test.default}
\alias{dunn.test.formula}
\title{Dunn test for multiple comparisons using rank sums}
\usage{
dunn.test(x, ...)

\method{dunn.test}{default}(
  x,
  g,
  compare.all = TRUE,
  baseline.group = ifelse(missing(g), NULL, levels(factor(g))[1]),
  contrast = NULL,
  p.adj.method = p.adjust.methods,
  alpha = 0.05
)

\method{dunn.test}{formula}(formula, data, ...)
}
\arguments{
\item{x}{a numeric vector of data values, or a list of numeric data vectors.
Non-numeric elements of a list will be coerced, with a warning.}

\item{g}{a vector or factor object giving the group for the corresponding
elements of x. Ignored with a warning if x is a list.}

\item{compare.all}{if \code{TRUE} (default), will consider all possible
between-group comparisons}

\item{baseline.group}{if \code{compare.all = FALSE}, will compare this group to
all others.}

\item{contrast}{if \code{compare.all = FALSE} and \code{baseline.group = NULL}, matrix
specifying groups to compare. See Details.}

\item{p.adj.method}{method for adjusting P values to account for multiple
comparisons. Default is the Bonferroni correction, following the original
manuscript. See \code{\link[=p.adjust]{p.adjust()}}.}

\item{alpha}{level at which the null hypothesis is rejected. See Details.}

\item{formula}{a formula of the form response ~ group where response gives
the data values and group a vector or factor of the corresponding groups.}
}
\value{
Object of class \code{dunn.htest}. If the ensemble test is passed, the
result will also have class \code{pairwise.htest}. If the ensemble test is not
passed and the between-group comparisons are not run, the result will also
have class \code{htest}.

This design provides for a specially-formatted \code{print} method (\link{print.dunn.htest}), as
well as downstream compatibility with methods expecting \code{htest} or \code{pairwise.htest}
methods as input (such as \code{\link[broom:reexports]{broom::tidy()}}). It also holds additional information
not provided for by \code{pairwise.htest}.

\code{dunn.htest} is a list containing the following elements:
\itemize{
\item \code{method} Name of the procedure
\item \code{data.name} Input data
\item \code{p.value} Matrix of adjusted P values. \code{NxN}, where \code{N} is the number of groups. Or a single P value if only the ensemble test is run.
\item \code{p.adjust.method} Adjustment method for multiple comparisons
\item \code{p.unadjust} Matrix of unadjusted P values
\item \code{statistic} Matrix of Z scores
\item \code{statistic.name} By analogy to the \code{htest} class
\item \code{contrast.table} \code{data.frame} containing the above information in a potentially
more useful format, with each row reporesenting one between-group comparison.
\item \code{alpha} Level for rejecting the null hypothesis
\item \code{pass.pre.test} \code{TRUE} if the ensemble test was passed
\item \code{pre.test} An \code{htest} object containing the results for the Kruskal-Wallis ensemble test.
Included only if the ensemble test is passed. If the ensemble test is not passed it is returned directly.
}
}
\description{
Non-parametric testing procedure for identifying differences between multiple
groups in a population. Implementation of Dunn JO (1964). Multiple
Comparisons Using Rank Sums, Technometrics, 6:3, 241-252. doi: (10.1080/00401706.1964.10490181)\link{https://doi.org/10.1080/00401706.1964.10490181}
}
\details{
This package implements Dunn's procedure for identifying differences
between multiple groups within a population. It is a two-step procedure,
first deciding if all groups come from identically-distributed populations,
then, if not, proceeding to post-hoc comparisons between individual groups.
The procedure compares the average ranks of values in each group. The
first-step ensemble test is the Kruskal-Wallis rank sum test
(\code{\link[=kruskal.test]{kruskal.test()}}). The post-hoc tests are similar to Wilcoxan rank sum
tests (\code{\link[=pairwise.wilcox.test]{pairwise.wilcox.test()}}), but the ranks of values within the larger sample
are maintained. The procedure includes an adjustment for ties.

Essentially, the Dunn test is a non-parametric analog of a one-way ANOVA
ensemble test followed by post-hoc tests such as Tukey. It has become
commonly used in biomedical research, likely because it is included in
popular statistics and graphing software packages (ex.:
\link{https://www.graphpad.com/guides/prism/latest/statistics/how_the_kruskal-wallis_test_works.htm}).

In principle, any subset of possible between-group comparisons may be
considered. By default, all possible comparisons are made. Alternatively, a
baseline group may be compared to all others by setting \code{compare.all = FALSE} and indicating the baseline group with \code{baseline.group}. To specify
an arbitrary set of comparisons, also set \code{baseline.group = NULL} and
identify desired between-group comparisons as an \code{NxN} \code{logical} matrix,
where \code{N} is the number of groups. Matrix row and column names should be
set to the group levels. A full set of comparisons is indicated by the
lower triangle (\code{\link[=lower.tri]{lower.tri()}}); the upper triangle gives the same result,
but ordered differently. Note that comparing a group to itself (diagonal
elements) or double-counting comparisons will adversely affect the results.

Dunn applies the Bonferroni correction to control the family-wise error
rate, and that is the default here. This approach can be overridden using
the \code{p.adj.method} parameter to specify any of the methods supported by
\code{\link[=p.adjust]{p.adjust()}}.

The post-hoc between-group inferences depend on the determination of the
preceding ensemble test. Therefore, if a Kruskal-Wallis fails to reject the
null hypothesis that all groups are drawn from identically-distributed
populations (\code{P >= alpha}), between-group comparisons will not run, and
only the ensemble test is reported. To override this behavior and force the
post-hoc tests to run, increase \code{alpha} above its default value of .05.
}
\examples{
dunn.test(mpg ~ cyl, data = mtcars)

}
\references{
Dunn JO (1964). Multiple Comparisons Using Rank Sums, Technometrics, 6:3, 241-252.  doi: (10.1080/00401706.1964.10490181)\link{https://doi.org/10.1080/00401706.1964.10490181}
}
