% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distill.R
\name{get_distill_source}
\alias{get_distill_source}
\title{Get the source URL for a Distill post}
\usage{
get_distill_source(
  repo = "taylordunn/tdunn",
  branch = "main",
  collection = "_posts",
  date = NULL,
  slug = NULL,
  html_tag = TRUE,
  html_label = "Source code"
)
}
\arguments{
\item{repo}{The GitHub repository, "user/repo-name".}

\item{branch}{Name of the GitHub branch, default "main".}

\item{collection}{Collection to which the post belongs to, default "_posts".}

\item{date}{The date of the post. Typically this will be defined as a
RMarkdown parameter via \code{params$date}.}

\item{slug}{The URL slug of the post. Typically this will be defined as a
RMarkdown parameter via \code{params$slug}.}

\item{html_tag}{Logical. If \code{TRUE}, returns a HTML hyperlink tag.}

\item{html_label}{If returning a HTML hyperlink, the text label to display.}
}
\value{
A string of the source code URL, or a HTML hyperlink tag.
}
\description{
This is a utility function that I use to point to source code on GitHub for
my Distill posts, located at https://tdunn.ca/.
Credit for the idea goes to Danielle Navarro (https://blog.djnavarro.net/).
}
\examples{

date <- "2021-05-18"
slug <- "tidytuesday-week-21"

get_distill_source(date = date, slug = slug, html_tag = FALSE)
get_distill_source(date = date, slug = slug, html_tag = TRUE)
get_distill_source(date = date, slug = slug, html_tag = TRUE,
                   html_label = "Custom label")
}
