% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dunn.test.R
\name{print.dunn.htest}
\alias{print.dunn.htest}
\title{Print method for Dunn test results}
\usage{
\method{print}{dunn.htest}(x, digits = getOption("digits"), prefix = "\\t", ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Print method for Dunn test results
}
