% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_axis.R
\name{remove_axis}
\alias{remove_axis}
\title{Removes one or both axes of a \code{ggplot2} plot}
\usage{
remove_axis(axis = "y")
}
\arguments{
\item{axis}{Specifies the axis to remove. One of "x", "y", "xy". Default is
"y" because my most common use case is for density plots.}
}
\description{
A helper function to set axis.line, axis.ticks, axis.text and axis.title
elements to \code{element_blank()} for the specified axis.
}
\examples{
library(ggplot2)
library(dplyr)
library(palmerpenguins)

p <- penguins \%>\%
  filter(!is.na(flipper_length_mm)) \%>\%
  ggplot(aes(x = flipper_length_mm)) +
  geom_density(aes(fill = species), alpha = 0.5) +
  scale_fill_brewer(palette = "Set1") +
  scale_y_continuous(expand = expansion(c(0, 0.1))) +
  labs(x = "Flipper length (mm)")
p + theme_td() +
  remove_axis("y")

}
