% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{td_colors}
\alias{td_colors}
\title{Colors and palettes}
\format{
An object of class \code{list} of length 4.
}
\usage{
td_colors
}
\description{
This list contains some nice colors and full palettes.
}
\examples{
library(scales)
show_col(unlist(td_colors$nice), ncol = 5)
show_col(td_colors$pastel6)
show_col(td_colors$div5)
}
\keyword{datasets}
