% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\docType{data}
\name{td_fonts}
\alias{td_fonts}
\title{Fonts}
\format{
An object of class \code{list} of length 2.
}
\usage{
td_fonts
}
\description{
This list contains some nice serif and sans serif fonts.
}
\keyword{datasets}
