% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{td_pal}
\alias{td_pal}
\title{TD palettes}
\usage{
td_pal(palette = "pastel6", type = "discrete", reverse = FALSE)
}
\arguments{
\item{palette}{Character name of palette:
\itemize{
\item "pastel6" (default)
\item "div5"
}}

\item{type}{One of "discrete" (default) or "continuous"}

\item{reverse}{Boolean to reverse palette order.}
}
\description{
Returns a function for a discrete TD color palette, or for interpolating a
continuous one via \code{grDevices::colorRampPalette()}.
}
\examples{
library(scales)
show_col(td_pal("pastel6")(6))
show_col(td_pal("pastel6", reverse = TRUE)(3))
show_col(td_pal("div5")(5))
show_col(td_pal("div5")(10))
}
