% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_td}
\alias{theme_td}
\title{The default TD theme for \code{ggplot2}}
\usage{
theme_td(
  base_size = 14,
  base_family = "Roboto Condensed",
  base_line_size = base_size/24
)
}
\arguments{
\item{base_size}{base font size in pts.}

\item{base_family}{base font family.}

\item{base_line_size}{base size for line elements.}
}
\value{
The complete theme.
}
\description{
A simple plotting theme with a half-open frame and no gridlines.
Heavily inspired by \code{cowplot::theme_half_open()}.
}
\examples{
library(ggplot2)
library(palmerpenguins)
library(dplyr)

# Load the default font (Roboto Condensed)
extrafont::loadfonts(quiet = TRUE)

p <- penguins \%>\%
  filter(!is.na(bill_length_mm)) \%>\%
  ggplot(aes(x = bill_length_mm, y = bill_depth_mm, color = species)) +
  geom_point(aes(shape = species), size = 3, alpha = 0.7) +
  geom_smooth(method = "lm", formula = "y ~ x", se = FALSE) +
  scale_color_brewer(palette = "Set1") +
  labs(title = "Penguin bill dimensions",
       subtitle = "Bill length and depth for different penguin species",
       x = "Bill length (mm)", y = "Bill depth (mm)",
       color = "Penguin species", shape = "Penguin species",
       caption = "Data from the palmerpenguins package") +
  facet_wrap(~species, nrow = 1)

p + theme_td()

}
