% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_td_grid}
\alias{theme_td_grid}
\title{Add the grid theme from dunnr to a ggplot}
\usage{
theme_td_grid(
  base_size = 12,
  base_family = "Droid Serif",
  base_line_size = base_size/24
)
}
\arguments{
\item{base_size}{base font size in pts.}

\item{base_family}{base font family.}

\item{base_line_size}{base size for line elements.}
}
\value{
The complete theme.
}
\description{
This theme is just \code{\link[=theme_bw]{theme_bw()}} with some slight tweaks, like softer colors
and Droid Serif as the default font.
}
\examples{
library(ggplot2)
library(palmerpenguins)
library(dplyr)

# Load the default font (Droid Serif)
extrafont::loadfonts(quiet = TRUE)

p <- penguins \%>\%
  filter(!is.na(flipper_length_mm)) \%>\%
  ggplot(aes(x = flipper_length_mm, y = species)) +
  geom_violin(aes(fill = species), show.legend = FALSE) +
  labs(x = "Flipper length (mm)", y = NULL) +
  facet_wrap(~year) +
  scale_fill_brewer(palette = "Set1") +
  scale_x_continuous(breaks = seq(180, 220, 20))

p + theme_td_grid()
}
